OLED Arduino LOGO Splash Screen
This Sketch adds the ARDUINO Logo as a Splash Screen for 5 seconds on a 128x64 OLED Display, then the display will go blank allowing you to add whatever code you wish to LOOP in the Void Loop.  If you wish to see the Splash Screen again press the Arduino's RESET Button.